changed data type for pgsql in heroku part 7 with ::decimal

+++ database/migrations/2020_07_19_154942_create_farmer_sales_view.php
@@ -20,11 +20,11 @@ class CreateFarmerSalesView extends Migration
             posts.user_id AS user_id,
             posts.crop_name AS crop_name,
             posts.created_at AS created_at,
-            SUM(posts.fixed_quantity) AS totalFixedQty,
-            SUM(posts.crop_quantity) AS totalAvailableQty,
+            SUM(posts.fixed_quantity::decimal) AS totalFixedQty,
+            SUM(posts.crop_quantity::decimal) AS totalAvailableQty,
             posts.earnings AS earnings,
-            SUM(posts.kilogram_sold) AS totalKgSold,
-            AVG(ROUND(posts.percentage_sold_before_harvest,
+            SUM(posts.kilogram_sold::decimal) AS totalKgSold,
+            AVG(ROUND(posts.percentage_sold_before_harvest::decimal,
                     1)) AS totalPercentage
         FROM
             posts
